Unison.Util.Nametree

Nametree

data Nametree a

A nametree has a value, and a collection of children nametrees keyed by name segment.

Constructors

Nametree

Fields

value :: !a

children :: !(Map NameSegment (Nametree a))

traverseNametreeWithName :: Applicative f => ([NameSegment] -> a -> f b) -> Nametree a -> f (Nametree b)

Traverse over a nametree, with access to the list of name segments (in reverse order) leading to each value.

unfoldNametree :: (a -> (b, Map NameSegment a)) -> a -> Nametree b

Build a nametree from a seed value.

Flattening and unflattening

flattenNametree :: forall a b. Ord b => (a -> Map NameSegment b) -> Nametree a -> BiMultimap b Name

flattenNametree organizes a nametree like

"foo" = #foo
"foo": {
  "bar" = #bar
  "bar": {
    "baz" = #baz
  }
}

into an equivalent-but-flatter association between names and definitions, like

{
  "foo" = #bar,
  "foo.bar" = #bar,
  "foo.bar.baz" = #baz
}

unflattenNametree :: Ord a => BiMultimap a Name -> Nametree (Map NameSegment a)

unflattenNametree organizes an association between names and definitions like

{
  "foo" = #bar,
  "foo.bar" = #bar,
  "foo.bar.baz" = #baz
}

into an equivalent-but-less-flat nametree, like

"foo" = #foo
"foo": {
  "bar" = #bar
  "bar": {
    "baz" = #baz
  }
}

Definitions

data Defns terms types

Definitions (terms and types) in a namespace.

FIXME this doesn't belong in this module

Constructors

Defns

Fields

terms :: !terms

types :: !types

mapDefns :: (a -> b) -> Defns a a -> Defns b b

bimapDefns :: (terms -> terms') -> (types -> types') -> Defns terms types -> Defns terms' types'